chore: remove legacy bundle
