+++ analyzer.html
@@ -24,7 +24,10 @@
     <link rel="stylesheet" href="style.css">
     <!-- Teachable Machine Image Library -->
     <script src="https://cdn.jsdelivr.net/npm/@teachablemachine/image@0.8/dist/teachablemachine-image.min.js"></script>
-    <script src="script.js" defer></script>
+    <script type="module" src="nav.js"></script>
+    <script type="module" src="upload.js"></script>
+    <script type="module" src="camera.js"></script>
+    <script type="module" src="analyzer.js"></script>
 </head>
 <body class="bg-gradient-to-br from-orange-50 to-red-50 min-h-screen flex flex-col">
     <!-- Navigation -->
@@ -45,7 +48,7 @@
                     <a href="about.html" class="nav-btn text-gray-700 hover:text-thai-red font-medium">เกี่ยวกับเรา</a>
                 </div>
                 <div class="md:hidden">
-                    <button onclick="toggleMobileMenu()" class="text-gray-700">
+                    <button onclick="nav.toggleMobileMenu()" class="text-gray-700">
                         <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                             <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M4 6h16M4 12h16M4 18h16"></path>
                         </svg>
@@ -107,7 +110,7 @@
                         <img id="previewImg" class="max-w-full max-h-64 mx-auto rounded-lg shadow-md mb-4">
                         <p class="text-gray-600 mb-4">ภาพที่เลือก</p>
                         <div class="flex flex-col sm:flex-row justify-center gap-4 mb-4">
-                            <button onclick="analyzeImage()" class="bg-thai-red hover:bg-red-700 text-white px-8 py-3 rounded-lg font-semibold transition-colors">
+                            <button onclick="analyzer.analyzeImage()" class="bg-thai-red hover:bg-red-700 text-white px-8 py-3 rounded-lg font-semibold transition-colors">
                                 🧠 วิเคราะห์ลวดลาย
                             </button>
                             <button id="removeImgBtn" class="bg-gray-300 hover:bg-gray-400 text-gray-800 px-6 py-3 rounded-lg font-semibold transition-colors">
@@ -147,7 +150,7 @@
                                 <p id="fabricDescription" class="text-gray-600 text-sm leading-relaxed">
                                     ผ้ามัดหมี่เป็นผ้าทอพื้นเมืองของภาคตะวันออกเฉียงเหนือ มีเทคนิคการมัดด้ายก่อนย้อมสี ทำให้เกิดลวดลายที่มีเอกลักษณ์เฉพาะ
                                 </p>
-                                <button onclick="showFabricDetails()" class="mt-4 bg-thai-gold hover:bg-yellow-500 text-thai-blue px-4 py-2 rounded-lg font-medium transition-colors">
+                                <button onclick="analyzer.showFabricDetails()" class="mt-4 bg-thai-gold hover:bg-yellow-500 text-thai-blue px-4 py-2 rounded-lg font-medium transition-colors">
                                     📚 เรียนรู้เพิ่มเติม
                                 </button>
                             </div>
